CellLabeler:

CellLabeler:

# S3 method for CellLabeler
celllabeler(
  object,
  features = NULL,
  markers = NULL,
  sample.var = "sample",
  cluster.var = "cluster",
  similar.gene = 20,
  similar.pct = 0.8,
  lfc = 0.5,
  max.genes = 100,
  min.ccells = 20,
  up.thr = 0.9,
  mod = "combine",
  num.core = 1,
  verbose = TRUE,
  ...
)

Arguments

object

A CellLabeler object

features

Feature names. Default is NULL, then all features will be used.

sample.var

Column names of sample information in metadata.

cluster.var

Column names of cluster information in metadata.

similar.gene

Integer value as the number of top expressed genes in considering combining clusters before DE test

similar.pct

Numeric value in 0,1 as the threshold to combine clusters before DE test when either two cluster have enough overlapped highly expressed genes. Default is 0.8.

lfc

Limit testing to genes which show the maximum on average X-fold difference (log-scale) between the interested clusters and the rest others. Default is 0.5.

max.genes

Maximum number of highly expressed genes in each cluster required. Default is 100.

min.ccells

Minimum number of cells in each cluster required. Default is 10.

up.thr

Numberic value in 0,1 as the proportion of the remaining clusters in which a gene is upregulated in the interested group compared with it. Highter, the more strict. Default 0.9.

mod

Character as "combine" or "all", specifying if combining similar clusters or not before DE test. Default is "combine"

num.core

Number of cores in multi-threaded running. Default is 1.

Value

object A CellLabeler object

Author

Jin Ning, Shiquan Sun